% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbam.R
\name{getbam_align}
\alias{getbam_align}
\title{Get Bam - align}
\usage{
getbam_align(
  fastq1,
  fastq2,
  reference,
  SNP.database,
  samtools.dir,
  picard.dir,
  bedtools.dir,
  GATK.dir,
  bwa.dir,
  sample.id,
  output.dir,
  java.dir = "java"
)
}
\description{
Map raw reads from fastq (gzip or not) files to the reference genome.
}
