% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbam.R
\name{getbam_align_after_merge}
\alias{getbam_align_after_merge}
\title{Get Bam - align after merge}
\usage{
getbam_align_after_merge(
  fastq1,
  fastq2,
  reference,
  SNP.database,
  samtools.dir,
  picard.dir,
  bedtools.dir,
  GATK.dir,
  bwa.dir,
  flash.dir,
  sample.id,
  output.dir,
  java.dir = "java"
)
}
\description{
First merge the overlapping read mates in cfDNA sequencing data and
then map raw reads from fastq (gzip or not) files to the reference genome
}
