% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{parameter_recommend}
\alias{parameter_recommend}
\title{Parameter Recommendation}
\usage{
parameter_recommend(
  plasma.unmerged,
  normal,
  plasma.merged.extendedFrags,
  plasma.merge.notCombined,
  target.bed,
  reference,
  SNP.database,
  samtools.dir,
  sample.id,
  roughly_estimated_tf = FALSE,
  python.dir
)
}
\arguments{
\item{plasma.unmerged}{a BAM file of the cfDNA sequencing data.}

\item{normal}{a BAM file of normal sample sequencing data.}

\item{plasma.merged.extendedFrags}{a BAM file of extendedFrags.}

\item{plasma.merge.notCombined}{a BAM file of reads not combined.}

\item{target.bed}{a BED file of target regions.}

\item{reference}{a FASTA file of reference genome.}

\item{SNP.database}{a VCF file of SNP database.}

\item{samtools.dir}{the path to an executable Samtools.}

\item{sample.id}{a sample name.}

\item{roughly_estimated_tf}{if calculate the roughly estimated tumor fraction.}

\item{python.dir}{the path to Python.}
}
\description{
Report the range of sequencing depth, the roughly estimated tumor fraction, and the detection limit range.
}
\examples{
## input files
demo.dir <- '/Users/huran/Desktop/cfSNV_development/demo/'
plasma.unmerged <- paste0(demo.dir, 'plasma.recal.bam')
normal <- paste0(demo.dir, 'normal.recal.bam')
plasma.merged.extendedFrags <- paste0(demo.dir, 'plasma.extendedFrags.recal.bam')
plasma.merge.notCombined <- paste0(demo.dir, 'plasma.notCombined.recal.bam')
target.bed <- paste0(demo.dir, 'example_target_regions.bed')
reference <- paste0(demo.dir, 'chr22.fa')
SNP.database <- paste0(demo.dir, 'chr22_dbSNP.vcf')
samtools.dir <- '/usr/local/bin/samtools'
sample.id <- '1st'
python.dir <- '/usr/local/bin/python3'


parameter_recommend(
  plasma.unmerged, normal,
  plasma.merged.extendedFrags, plasma.merge.notCombined,
  target.bed, reference, SNP.database, samtools.dir, sample.id, python.dir=python.dir
)
}
