% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{variant_calling}
\alias{variant_calling}
\title{Variant Calling}
\usage{
variant_calling(
  plasma.unmerged,
  normal,
  plasma.merged.extendedFrags,
  plasma.merge.notCombined,
  target.bed,
  reference,
  SNP.database,
  samtools.dir,
  picard.dir,
  bedtools.dir,
  sample.id,
  MIN_HOLD_SUPPORT_COUNT = 12,
  MIN_PASS_SUPPORT_COUNT = 5,
  java.dir = "java",
  python.dir
)
}
\arguments{
\item{plasma.unmerged}{a BAM file of the cfDNA sequencing data.}

\item{normal}{a BAM file of normal sample sequencing data.}

\item{plasma.merged.extendedFrags}{a BAM file of extendedFrags.}

\item{plasma.merge.notCombined}{a BAM file of reads not combined.}

\item{target.bed}{a BED file of target regions.}

\item{reference}{a .fa file of reference genome.}

\item{SNP.database}{a .vcf of SNP database.}

\item{samtools.dir}{the path to an executable Samtools.}

\item{picard.dir}{the path to picard.jar.}

\item{bedtools.dir}{the path to an executable BEDTools.}

\item{sample.id}{the name of sample.}

\item{MIN_HOLD_SUPPORT_COUNT}{default is 12.}

\item{MIN_PASS_SUPPORT_COUNT}{default is 5.}

\item{java.dir}{the path to java, default is "java".}

\item{python.dir}{the path to Python.}
}
\value{
a variant list and a tumor fraction
}
\description{
This function allows you to call somatic mutations.
}
\examples{
## input files
demo.dir <- '/Users/huran/Desktop/cfSNV_development/demo/'
plasma.unmerged <- paste0(demo.dir, 'plasma.recal.bam')
normal <- paste0(demo.dir, 'normal.recal.bam')
plasma.merged.extendedFrags <- paste0(demo.dir, 'plasma.extendedFrags.recal.bam')
plasma.merge.notCombined <- paste0(demo.dir, 'plasma.notCombined.recal.bam')
target.bed <- paste0(demo.dir, 'example_target_regions.bed')
reference <- paste0(demo.dir, 'chr22.fa')
SNP.database <- paste0(demo.dir, 'chr22_dbSNP.vcf')
samtools.dir <- '/usr/local/bin/samtools'
picard.dir <- '/usr/local/bin/picard.jar'
bedtools.dir <- '/usr/local/bin/bedtools2/bin/bedtools'
sample.id <- '1st'
MIN_HOLD_SUPPORT_COUNT <- 9
MIN_PASS_SUPPORT_COUNT <- 3
python.dir <- '/usr/local/bin/python3'

results <- variant_calling(
  plasma.unmerged, normal, plasma.merged.extendedFrags, plasma.merge.notCombined,
  target.bed, reference, SNP.database, samtools.dir, picard.dir, bedtools.dir,
  sample.id, MIN_HOLD_SUPPORT_COUNT, MIN_PASS_SUPPORT_COUNT, python.dir=python.dir
)
}
